requirements.txt
lighter-sdk
python-dotenv
pandas
numpy
aiohttp
asyncio
loguru
